Updated the Investor Service Spec
At some point, the Investor Already Exists response was amended to include an
investor ID, however the tests were not updated correctly to reflect the new
logic.

The logic in place would suggest that if we get an Investor Already Exists
RDS code, then the data payload should include an investor ID - otherwise we
throw an error.

I've amended the tests to reflect this logic.